requirements.txt
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-Limiter==4.1.1
Flask-Migrate==4.1.0
SQLAlchemy==2.0.48
Werkzeug==3.1.6
python-dotenv==1.2.2
email-validator==2.3.0
pytz==2026.1.post1
